# OpenAPI.yaml
openapi: 3.0.3
info:
  title: API Blog Comments
  version: "1.0"
  description: >
    HTTP API for managing blog posts, comments, and authenticated users.
    The contract reflects a production-oriented baseline with real persistence,
    JWT-based authentication, ownership rules, and interactive documentation support.
servers:
  - url: http://localhost:5000
    description: Local execution endpoint
...
paths:
  /api/posts:
    get:
      tags: [Posts]
      summary: List blog posts
      description: Returns blog posts in summary format using pagination.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          description: Requested page number.
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          description: Number of items returned per page.
      responses:
        '200':
          description: Post summary list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBlogPostSummaryResult'
    post:
      tags: [Posts]
      summary: Create blog post
      description: Creates a new blog post owned by the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogPost'
            example:
              title: Production-ready API baseline
              content: This post describes the architectural goals of the project.
      responses:
        '201':
          description: Blog post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '400':
          description: Invalid request payload
        '401':
          description: Unauthorized
  /api/posts/{id}:
    get:
      tags: [Posts]
      summary: Get blog post by identifier
      description: Returns a single blog post together with its comments.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Blog post returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Not found
    put:
      tags: [Posts]
      summary: Update blog post
      description: Updates a blog post when the caller is the owner or an administrator.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogPost'
            example:
              title: Updated architectural baseline
              content: Revised content for the existing blog post.
      responses:
        '200':
          description: Blog post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '400':
          description: Invalid request payload
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags: [Posts]
      summary: Delete blog post
      description: Removes a blog post when the caller is the owner or an administrator.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: Blog post deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /api/posts/{id}/comments:
    get:
      tags: [Comments]
      summary: List post comments
      description: Returns comments associated with a given blog post using pagination.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          description: Requested page number.
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          description: Number of items returned per page.
      responses:
        '200':
          description: Comment list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCommentResult'
        '404':
          description: Post not found
    post:
      tags: [Comments]
      summary: Create comment
      description: Creates a new comment associated with the specified blog post.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComment'
            example:
              text: This is a relevant comment about the published post.
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid request payload
        '401':
          description: Unauthorized
        '404':
          description: Post not found
  /api/posts/{id}/comments/{commentId}:
    get:
      tags: [Comments]
      summary: Get comment by identifier
      description: Returns a single comment associated with the specified blog post.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: path
          name: commentId
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Comment returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Comment not found
    put:
      tags: [Comments]
      summary: Update comment
      description: Updates a comment when the caller is the owner or an administrator.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: path
          name: commentId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComment'
            example:
              text: Updated comment text.
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid request payload
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Comment not found
    delete:
      tags: [Comments]
      summary: Delete comment
      description: Removes a comment when the caller is the owner or an administrator.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: path
          name: commentId
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Comment not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BlogPost:
      type: object
      required:
        - id
        - title
        - content
        - createdByUserId
        - createdByUsername
        - createdAt
        - updatedAt
        - comments
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: API Blog Comments
        content:
          type: string
          example: Reference implementation for a production-oriented API baseline.
        createdByUserId:
          type: integer
          example: 10
        createdByUsername:
          type: string
          example: author-user
        createdAt:
          type: string
          format: date-time
          example: 2026-04-04T18:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-04-04T18:30:00Z
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    BlogPostSummary:
      type: object
      required:
        - id
        - title
        - createdByUserId
        - createdByUsername
        - createdAt
        - updatedAt
        - commentCount
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: API Blog Comments
        createdByUserId:
          type: integer
          example: 10
        createdByUsername:
          type: string
          example: author-user
        createdAt:
          type: string
          format: date-time
          example: 2026-04-04T18:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-04-04T18:30:00Z
        commentCount:
          type: integer
          example: 3
    PagedBlogPostSummaryResult:
      type: object
      required:
        - items
        - page
        - pageSize
        - totalCount
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BlogPostSummary'
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalCount:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
    CreateBlogPost:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          maxLength: 200
          example: Production-ready API baseline
        content:
          type: string
          example: This post explains the technical goals of the project.
    Comment:
      type: object
      required:
        - id
        - text
        - createdByUserId
        - createdByUsername
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          example: 1
        text:
          type: string
          example: This is a relevant technical observation.
        createdByUserId:
          type: integer
          example: 10
        createdByUsername:
          type: string
          example: reviewer-user
        createdAt:
          type: string
          format: date-time
          example: 2026-04-04T18:05:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-04-04T18:10:00Z
    PagedCommentResult:
      type: object
      required:
        - items
        - page
        - pageSize
        - totalCount
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalCount:
          type: integer
          example: 12
        totalPages:
          type: integer
          example: 1
    CreateComment:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          maxLength: 500
          example: This is a relevant comment about the post.
...